Allow disabling basic authentication via shoot spec
