+++ charts/seed-monitoring/charts/core/charts/prometheus/templates/prometheus.yaml
@@ -142,6 +142,14 @@ spec:
         env:
         - name: OPENVPN_PORT
           value: "4314"
+        - name: APISERVER_AUTH_MODE
+          value: client-cert
+        - name: APISERVER_AUTH_MODE_CLIENT_CERT_CA
+          value: /srv/secrets/vpn-seed/ca.crt
+        - name: APISERVER_AUTH_MODE_CLIENT_CERT_CRT
+          value: /srv/secrets/vpn-seed/tls.crt
+        - name: APISERVER_AUTH_MODE_CLIENT_CERT_KEY
+          value: /srv/secrets/vpn-seed/tls.key
         - name: SERVICE_NETWORK
           value: {{ .Values.networks.services }}
         - name: POD_NETWORK
@@ -171,8 +179,6 @@ spec:
           name: vpn-seed
         - mountPath: /srv/secrets/tlsauth
           name: vpn-seed-tlsauth
-        - mountPath: /srv/auth
-          name: kube-apiserver-basic-auth
       - name: blackbox-exporter
         image: {{ index .Values.images "blackbox-exporter" }}
         args:
@@ -243,9 +249,6 @@ spec:
       - name: etcd-client-tls
         secret:
           secretName: etcd-client-tls
-      - name: kube-apiserver-basic-auth
-        secret:
-          secretName: kube-apiserver-basic-auth
       - name: prometheus-kubeconfig
         secret:
           secretName: prometheus
