+++ pkg/operation/botanist/controlplane.go
@@ -27,15 +27,11 @@ import (
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
 	gardenv1beta1 "github.com/gardener/gardener/pkg/apis/garden/v1beta1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
-	controllermanagerfeatures "github.com/gardener/gardener/pkg/controllermanager/features"
-	"github.com/gardener/gardener/pkg/features"
 	"github.com/gardener/gardener/pkg/migration"
 	"github.com/gardener/gardener/pkg/operation/common"
-	"github.com/gardener/gardener/pkg/utils"
 	kutil "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/kubernetes/health"
 	"github.com/gardener/gardener/pkg/utils/retry"
-	"github.com/gardener/gardener/pkg/utils/secrets"
 
 	appsv1 "k8s.io/api/apps/v1"
 	autoscalingv2beta1 "k8s.io/api/autoscaling/v2beta1"
@@ -255,323 +251,6 @@ func (b *Botanist) DeleteClusterAutoscaler(ctx context.Context) error {
 		},
 	}
 	return client.IgnoreNotFound(b.K8sSeedClient.Client().Delete(ctx, deploy, kubernetes.DefaultDeleteOptionFuncs...))
-}
-
-// DeploySeedMonitoring will install the Helm release "seed-monitoring" in the Seed clusters. It comprises components
-// to monitor the Shoot cluster whose control plane runs in the Seed cluster.
-func (b *Botanist) DeploySeedMonitoring(ctx context.Context) error {
-	var (
-		credentials      = b.Secrets["monitoring-ingress-credentials"]
-		credentialsUsers = b.Secrets["monitoring-ingress-credentials-users"]
-		basicAuth        = utils.CreateSHA1Secret(credentials.Data[secrets.DataKeyUserName], credentials.Data[secrets.DataKeyPassword])
-		basicAuthUsers   = utils.CreateSHA1Secret(credentialsUsers.Data[secrets.DataKeyUserName], credentialsUsers.Data[secrets.DataKeyPassword])
-		prometheusHost   = b.ComputePrometheusHost()
-	)
-
-	var (
-		prometheusConfig = map[string]interface{}{
-			"kubernetesVersion": b.Shoot.Info.Spec.Kubernetes.Version,
-			"networks": map[string]interface{}{
-				"pods":     b.Shoot.GetPodNetwork(),
-				"services": b.Shoot.GetServiceNetwork(),
-				"nodes":    b.Shoot.GetNodeNetwork(),
-			},
-			"ingress": map[string]interface{}{
-				"basicAuthSecret": basicAuth,
-				"host":            prometheusHost,
-			},
-			"namespace": map[string]interface{}{
-				"uid": b.SeedNamespaceObject.UID,
-			},
-			"objectCount": b.Shoot.GetNodeCount(),
-			"podAnnotations": map[string]interface{}{
-				"checksum/secret-prometheus":                b.CheckSums["prometheus"],
-				"checksum/secret-kube-apiserver-basic-auth": b.CheckSums[common.BasicAuthSecretName],
-				"checksum/secret-vpn-seed":                  b.CheckSums["vpn-seed"],
-				"checksum/secret-vpn-seed-tlsauth":          b.CheckSums["vpn-seed-tlsauth"],
-			},
-			"replicas":           b.Shoot.GetReplicas(1),
-			"apiserverServiceIP": common.ComputeClusterIP(b.Shoot.GetServiceNetwork(), 1),
-			"seed": map[string]interface{}{
-				"apiserver": b.K8sSeedClient.RESTConfig().Host,
-				"region":    b.Seed.Info.Spec.Cloud.Region,
-				"profile":   b.Seed.Info.Spec.Cloud.Profile,
-			},
-			"rules": map[string]interface{}{
-				"optional": map[string]interface{}{
-					"cluster-autoscaler": map[string]interface{}{
-						"enabled": b.Shoot.WantsClusterAutoscaler,
-					},
-					"alertmanager": map[string]interface{}{
-						"enabled": b.Shoot.WantsAlertmanager,
-					},
-					"elasticsearch": map[string]interface{}{
-						"enabled": controllermanagerfeatures.FeatureGate.Enabled(features.Logging),
-					},
-				},
-			},
-			"shoot": map[string]interface{}{
-				"apiserver": fmt.Sprintf("https://%s", b.Shoot.InternalClusterDomain),
-				"provider":  b.Shoot.CloudProvider,
-			},
-			"ignoreAlerts": b.Shoot.IgnoreAlerts,
-		}
-		kubeStateMetricsSeedConfig = map[string]interface{}{
-			"replicas": b.Shoot.GetReplicas(1),
-		}
-		kubeStateMetricsShootConfig = map[string]interface{}{
-			"replicas": b.Shoot.GetReplicas(1),
-		}
-	)
-
-	prometheus, err := b.InjectSeedShootImages(prometheusConfig,
-		common.PrometheusImageName,
-		common.ConfigMapReloaderImageName,
-		common.VPNSeedImageName,
-		common.BlackboxExporterImageName,
-		common.AlpineIptablesImageName,
-	)
-	if err != nil {
-		return err
-	}
-	kubeStateMetricsSeed, err := b.InjectSeedShootImages(kubeStateMetricsSeedConfig, common.KubeStateMetricsImageName)
-	if err != nil {
-		return err
-	}
-	kubeStateMetricsShoot, err := b.InjectSeedShootImages(kubeStateMetricsShootConfig, common.KubeStateMetricsImageName)
-	if err != nil {
-		return err
-	}
-
-	coreValues := map[string]interface{}{
-		"global": map[string]interface{}{
-			"shootKubeVersion": map[string]interface{}{
-				"gitVersion": b.Shoot.Info.Spec.Kubernetes.Version,
-			},
-		},
-		"prometheus":               prometheus,
-		"kube-state-metrics-seed":  kubeStateMetricsSeed,
-		"kube-state-metrics-shoot": kubeStateMetricsShoot,
-	}
-
-	if err := b.ApplyChartSeed(filepath.Join(common.ChartPath, "seed-monitoring", "charts", "core"), b.Shoot.SeedNamespace, fmt.Sprintf("%s-monitoring", b.Shoot.SeedNamespace), nil, coreValues); err != nil {
-		return err
-	}
-
-	if err := b.deployGrafanaCharts("operators", basicAuth, common.GrafanaOperatorsPrefix); err != nil {
-		return err
-	}
-
-	if err := b.deployGrafanaCharts("users", basicAuthUsers, common.GrafanaUsersPrefix); err != nil {
-		return err
-	}
-
-	// Check if we want to deploy an alertmanager into the shoot namespace.
-	if b.Shoot.WantsAlertmanager {
-		var (
-			alertingSMTPKeys = b.GetSecretKeysOfRole(common.GardenRoleAlertingSMTP)
-			emailConfigs     = []map[string]interface{}{}
-			to, _            = b.Shoot.Info.Annotations[common.GardenOperatedBy]
-		)
-
-		for _, key := range alertingSMTPKeys {
-			secret := b.Secrets[key]
-			emailConfigs = append(emailConfigs, map[string]interface{}{
-				"to":            to,
-				"from":          string(secret.Data["from"]),
-				"smarthost":     string(secret.Data["smarthost"]),
-				"auth_username": string(secret.Data["auth_username"]),
-				"auth_identity": string(secret.Data["auth_identity"]),
-				"auth_password": string(secret.Data["auth_password"]),
-			})
-		}
-
-		alertManagerValues, err := b.InjectSeedShootImages(map[string]interface{}{
-			"ingress": map[string]interface{}{
-				"basicAuthSecret": basicAuth,
-				"host":            b.Seed.GetIngressFQDN("a", b.Shoot.Info.Name, b.Garden.Project.Name),
-			},
-			"replicas":     b.Shoot.GetReplicas(1),
-			"storage":      b.Seed.GetValidVolumeSize("1Gi"),
-			"emailConfigs": emailConfigs,
-		}, common.AlertManagerImageName, common.ConfigMapReloaderImageName)
-		if err != nil {
-			return err
-		}
-		if err := b.ApplyChartSeed(filepath.Join(common.ChartPath, "seed-monitoring", "charts", "alertmanager"), b.Shoot.SeedNamespace, fmt.Sprintf("%s-monitoring", b.Shoot.SeedNamespace), nil, alertManagerValues); err != nil {
-			return err
-		}
-	} else {
-		if err := common.DeleteAlertmanager(ctx, b.K8sSeedClient.Client(), b.Shoot.SeedNamespace); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (b *Botanist) deployGrafanaCharts(role, basicAuth, subDomain string) error {
-	values, err := b.InjectSeedShootImages(map[string]interface{}{
-		"ingress": map[string]interface{}{
-			"basicAuthSecret": basicAuth,
-			"host":            b.Seed.GetIngressFQDN(subDomain, b.Shoot.Info.Name, b.Garden.Project.Name),
-		},
-		"replicas": b.Shoot.GetReplicas(1),
-		"role":     role,
-	}, common.GrafanaImageName, common.BusyboxImageName)
-	if err != nil {
-		return err
-	}
-	return b.ApplyChartSeed(filepath.Join(common.ChartPath, "seed-monitoring", "charts", "grafana"), b.Shoot.SeedNamespace, fmt.Sprintf("%s-monitoring", b.Shoot.SeedNamespace), nil, values)
-}
-
-// DeleteSeedMonitoring will delete the monitoring stack from the Seed cluster to avoid phantom alerts
-// during the deletion process. More precisely, the Alertmanager and Prometheus StatefulSets will be
-// deleted.
-func (b *Botanist) DeleteSeedMonitoring(ctx context.Context) error {
-	alertManagerStatefulSet := &appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      common.AlertManagerStatefulSetName,
-			Namespace: b.Shoot.SeedNamespace,
-		},
-	}
-	if err := b.K8sSeedClient.Client().Delete(ctx, alertManagerStatefulSet, kubernetes.DefaultDeleteOptionFuncs...); client.IgnoreNotFound(err) != nil {
-		return err
-	}
-
-	prometheusStatefulSet := &appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      gardencorev1alpha1.StatefulSetNamePrometheus,
-			Namespace: b.Shoot.SeedNamespace,
-		},
-	}
-	return client.IgnoreNotFound(b.K8sSeedClient.Client().Delete(ctx, prometheusStatefulSet, kubernetes.DefaultDeleteOptionFuncs...))
-}
-
-// DeploySeedLogging will install the Helm release "seed-bootstrap/charts/elastic-kibana-curator" in the Seed clusters.
-func (b *Botanist) DeploySeedLogging(ctx context.Context) error {
-	if !controllermanagerfeatures.FeatureGate.Enabled(features.Logging) {
-		return common.DeleteLoggingStack(ctx, b.K8sSeedClient.Client(), b.Shoot.SeedNamespace)
-	}
-
-	var (
-		kibanaHost                             = b.Seed.GetIngressFQDN("k", b.Shoot.Info.Name, b.Garden.Project.Name)
-		kibanaCredentials                      = b.Secrets["kibana-logging-sg-credentials"]
-		kibanaUserIngressCredentialsSecretName = "logging-ingress-credentials-users"
-		sgKibanaUsername                       = kibanaCredentials.Data[secrets.DataKeyUserName]
-		sgKibanaPassword                       = kibanaCredentials.Data[secrets.DataKeyPassword]
-		sgKibanaPasswordHash                   = kibanaCredentials.Data[secrets.DataKeyPasswordBcryptHash]
-		basicAuth                              = utils.EncodeBase64([]byte(fmt.Sprintf("%s:%s", sgKibanaUsername, sgKibanaPassword)))
-
-		sgCuratorPassword     = b.Secrets["curator-sg-credentials"].Data[secrets.DataKeyPassword]
-		sgCuratorPasswordHash = b.Secrets["curator-sg-credentials"].Data[secrets.DataKeyPasswordBcryptHash]
-
-		sgUserPasswordHash  = b.Secrets[kibanaUserIngressCredentialsSecretName].Data[secrets.DataKeyPasswordBcryptHash]
-		sgAdminPasswordHash = b.Secrets[common.KibanaAdminIngressCredentialsSecretName].Data[secrets.DataKeyPasswordBcryptHash]
-
-		userIngressBasicAuth  = utils.CreateSHA1Secret(b.Secrets[kibanaUserIngressCredentialsSecretName].Data[secrets.DataKeyUserName], b.Secrets[kibanaUserIngressCredentialsSecretName].Data[secrets.DataKeyPassword])
-		adminIngressBasicAuth = utils.CreateSHA1Secret(b.Secrets[common.KibanaAdminIngressCredentialsSecretName].Data[secrets.DataKeyUserName], b.Secrets[common.KibanaAdminIngressCredentialsSecretName].Data[secrets.DataKeyPassword])
-		ingressBasicAuth      string
-
-		sgFluentdPasswordHash string
-	)
-
-	userIngressBasicAuthDecoded, err := utils.DecodeBase64(userIngressBasicAuth)
-	if err != nil {
-		return err
-	}
-
-	adminIngressBasicAuthDecoded, err := utils.DecodeBase64(adminIngressBasicAuth)
-	if err != nil {
-		return err
-	}
-
-	ingressBasicAuthDecoded := fmt.Sprintf("%s\n%s", string(userIngressBasicAuthDecoded), string(adminIngressBasicAuthDecoded))
-	ingressBasicAuth = utils.EncodeBase64([]byte(ingressBasicAuthDecoded))
-
-	images, err := b.InjectSeedSeedImages(map[string]interface{}{},
-		common.ElasticsearchImageName,
-		common.ElasticsearchMetricsExporterImageName,
-		common.ElasticsearchSearchguardImageName,
-		common.CuratorImageName,
-		common.KibanaImageName,
-		common.SearchguardImageName,
-		common.AlpineImageName,
-	)
-	if err != nil {
-		return err
-	}
-
-	ct := b.Shoot.Info.CreationTimestamp.Time
-
-	sgFluentdSecret := &corev1.Secret{}
-	if err = b.K8sSeedClient.Client().Get(ctx, kutil.Key(common.GardenNamespace, "fluentd-es-sg-credentials"), sgFluentdSecret); err != nil {
-		return err
-	}
-
-	sgFluentdPasswordHash = string(sgFluentdSecret.Data["bcryptPasswordHash"])
-
-	elasticKibanaCurator := map[string]interface{}{
-		"ingress": map[string]interface{}{
-			"host":            kibanaHost,
-			"basicAuthSecret": ingressBasicAuth,
-		},
-		"elasticsearch": map[string]interface{}{
-			"replicaCount": b.Shoot.GetReplicas(1),
-			"readinessProbe": map[string]interface{}{
-				"httpAuth": basicAuth,
-			},
-			"metricsExporter": map[string]interface{}{
-				"username": string(sgKibanaUsername),
-				"password": string(sgKibanaPassword),
-			},
-		},
-		"kibana": map[string]interface{}{
-			"replicaCount": b.Shoot.GetReplicas(1),
-			"sgUsername":   "kibanaserver",
-			"sgPassword":   string(sgKibanaPassword),
-		},
-		"curator": map[string]interface{}{
-			"hourly": map[string]interface{}{
-				"schedule": fmt.Sprintf("%d * * * *", ct.Minute()),
-				"suspend":  b.Shoot.HibernationEnabled,
-			},
-			"daily": map[string]interface{}{
-				"schedule": fmt.Sprintf("%d 0,6,12,18 * * *", ct.Minute()%54+5),
-				"suspend":  b.Shoot.HibernationEnabled,
-			},
-			"sgUsername": "curator",
-			"sgPassword": string(sgCuratorPassword),
-		},
-		"searchguard": map[string]interface{}{
-			"enabled":      true,
-			"replicaCount": b.Shoot.GetReplicas(1),
-			"annotations": map[string]interface{}{
-				"checksum/tls-secrets-server": b.CheckSums["elasticsearch-logging-server"],
-				"checksum/sg-admin-client":    b.CheckSums["sg-admin-client"],
-			},
-			"users": map[string]interface{}{
-				"fluentd": map[string]interface{}{
-					"hash": string(sgFluentdPasswordHash),
-				},
-				"kibanaserver": map[string]interface{}{
-					"hash": string(sgKibanaPasswordHash),
-				},
-				"curator": map[string]interface{}{
-					"hash": string(sgCuratorPasswordHash),
-				},
-				"user": map[string]interface{}{
-					"hash": string(sgUserPasswordHash),
-				},
-				"admin": map[string]interface{}{
-					"hash": string(sgAdminPasswordHash),
-				},
-			},
-		},
-		"global": images,
-	}
-
-	return b.ApplyChartSeed(filepath.Join(common.ChartPath, "seed-bootstrap", "charts", "elastic-kibana-curator"), b.Shoot.SeedNamespace, fmt.Sprintf("%s-logging", b.Shoot.SeedNamespace), nil, elasticKibanaCurator)
 }
 
 // DeployDependencyWatchdog deploys the dependency watchdog to the Shoot namespace in the Seed.
